Float values for integer 2

+++ calc.py
@@ -3,7 +3,7 @@
 # Hello world
 def compute(expression):
     num0, operator, num1 = expression.split(' ')
-    num0, num1 = float(num0), int(num1)
+    num0, num1 = float(num0), float(num1)
     if operator == '+':
         return num0 + num1
     elif operator == '-':
